Added an API integration + multiprocessing for OpenCV
Windows and Unix has their own version of the OpenCV API server. The Unix version makes use of multiprocessing since it supports forking processes. Windows does not resulting in slower processing speed since it re-launches a whole new python process for each thread. For this reason, multi-processing was not included in the Windows version.

Api integration between the client and the server is progressing too.