Fixed screenshot class on higher resolution displays
By default mss takes screenshot at the resolution of the display, which league displays on. League of Legends also by default displays at the maximum resolution. To make this work, the users needs to switch the display mode to "Fullscreen" and the resolution to "1920x1080". CyberLeague takes care of the rest by correctly rescaling/croping the image.